Rename: Lazy 함수를 모아놓은 `_index` 파일 추가
